% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/long2mat.R
\name{long2mat}
\alias{long2mat}
\title{Long format to Matrix format}
\usage{
long2mat(data_long, country_id, counter_id, contiguity,
  id_is_int = FALSE, W_style = NULL)
}
\arguments{
\item{data_long}{长格式的空间权重矩阵}

\item{country_id}{国家或地区的 ID 的列名}

\item{counter_id}{目标国家或地区的 ID 的列名}

\item{contiguity}{权重列的列名}

\item{id_is_int}{国家或地区的 ID 是否为整数型变量}

\item{W_style}{对空间权重做什么样的标准化处理，默认 \code{NULL} 为不做处理，参考包 \code{spdep}}
}
\value{
m_df 数据框格式的矩阵型空间权重矩阵，第一列代表 ID

check_list 原来字符型的 ID 与新生成的整数型 ID 的对照表
}
\description{
Long format to Matrix format
}
\examples{
long2mat(data_long = queen_long, country_id = "Country ISO3 Code",
         counter_id = "Counterpart Country ISO3 Code", contiguity = "Contiguity",
         id_is_int = FALSE, W_style = "W")
}
