% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{queen_long}
\alias{queen_long}
\title{Long-format spatial weight matrix of Queen}
\format{A data frame with 60516 rows and 7 columns.
\describe{
\item{Country Name}{Country name}
\item{Country ISO3 Code}{ISO3 country code corresponding to the country name}
\item{Country UN Code}{UN country code corresponding to the country name}
\item{Counterpart Country Name}{Counterpart country name}
\item{Counterpart Country ISO3 Code}{ISO3 counterpart country code corresponding to \
the counterpart country name}
\item{Counterpart Country UN Code}{UN counterpart country code corresponding to \
the counterpart country name}
\item{Contiguity}{Wether the country and the counterpart country are neighbors}
}}
\usage{
queen_long
}
\description{
Long-format spatial weight matrix of Queen
}
\keyword{internal}
