% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_stata_spmat.R
\name{write_stata_spmat}
\alias{write_stata_spmat}
\title{Write apatial weight matrix to a \code{.txt} file which can be read by Stata15}
\usage{
write_stata_spmat(long2mat_list, file_name, save_check_list = FALSE)
}
\arguments{
\item{long2mat_list}{output of \code{long2mat}}

\item{file_name}{just path and file name}

\item{save_check_list}{save ID check list as both \code{save_check_list = TRUE} and \code{id_is_int = FALSE}}
}
\description{
Write apatial weight matrix to a \code{.txt} file which can be read by Stata15
}
\examples{
\dontrun{
write_stata_spmat(long2mat_list = queen_list, file_name = "World_spmat", save_check_list = TRUE)
}
}
