% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{queen_long}
\alias{queen_long}
\title{Long-format of queen spatial weight matrix}
\format{A data frame with 60516 rows and 7 columns.
\describe{
\item{Country Name}{Country Name}
\item{Country ISO3 Code}{Country Name}
\item{Country UN Code}{Country Name}
\item{Counterpart Country Name}{Country Name}
\item{Counterpart Country ISO3 Code}{Country Name}
\item{Counterpart Country UN Code}{Country Name}
\item{Contiguity}{Country Name}
}}
\usage{
queen_long
}
\description{
Long-format of queen spatial weight matrix
}
\keyword{internal}
